2023 12 27 16:17 Connection Pool 오류 수정 완료 1. jsp에서 작성한 쿼리문들 전부 close 메소드 추가 완료 2. 포워드 분기마다 이루어졌던 close 메소드 하나로 통일 <-- 포워드가 이루어 졌어도 하위 코드는 	훑어보기 때문에 한번만 닫아줘도 됨.
